refactor: extract scenario command definitions
Move typed command definitions out of scenario.cpp and compute.hpp into
command_types.hpp. This separates the resolved scenario command model
from its consumers without changing behavior.
